+++ rana_qgis_plugin/loader.py
@@ -1,4 +1,5 @@
 import os
+from copy import deepcopy
 from functools import partial
 from pathlib import Path
 
@@ -66,6 +67,7 @@ class Loader(QObject):
     vector_style_finished = pyqtSignal()
     vector_style_failed = pyqtSignal(str)
     loading_cancelled = pyqtSignal()
+    download_results_cancelled = pyqtSignal()
     simulation_cancelled = pyqtSignal()
     simulation_started = pyqtSignal()
     simulation_started_failed = pyqtSignal()
@@ -119,7 +121,9 @@ class Loader(QObject):
 
     def on_file_download_finished(self, project, file, local_file_path: str):
         self.communication.clear_message_bar()
-        self.communication.bar_info(f"File(s) downloaded to: {local_file_path}")
+        self.communication.bar_info(
+            f"File(s) downloaded to: {local_file_path}", dur=240
+        )
         sender = self.sender()
         assert isinstance(sender, QThread)
         sender.wait()
@@ -164,7 +168,11 @@ class Loader(QObject):
 
     def on_file_download_progress(self, progress: int, file_name: str = ""):
         self.communication.progress_bar(
-            f"Downloading file {file_name}...", 0, 100, progress, clear_msg_bar=True
+            f"Generating/downloading file {file_name}...",
+            0,
+            100,
+            progress,
+            clear_msg_bar=True,
         )
         self.file_download_progress.emit(progress, file_name)
 
@@ -364,8 +372,17 @@ class Loader(QObject):
             self.communication.show_warn(
                 "Working directory not yet set, please configure this in the plugin settings."
             )
-
         descriptor = get_tenant_file_descriptor(file["descriptor_id"])
+        if (
+            not descriptor["meta"]
+            or not descriptor["meta"]["simulation"]
+            or not descriptor["meta"]["simulation"]["name"]
+        ):
+            self.communication.show_error(
+                "Scenario is corrupt; did you upload a zip directly?"
+            )
+            self.download_results_cancelled.emit()
+            return
         schematisation_name = descriptor["meta"]["schematisation"]["name"]
         schematisation_id = descriptor["meta"]["schematisation"]["id"]
         schematisation_version = descriptor["meta"]["schematisation"]["version"]
@@ -375,20 +392,25 @@ class Loader(QObject):
         target_folder = get_threedi_schematisation_simulation_results_folder(
             QgsSettings().value("threedi/working_dir"),
             schematisation_id,
-            schematisation_name,
+            schematisation_name.replace("/", "-").replace("\\", "-"),
             schematisation_version,
-            descriptor["meta"]["simulation"]["name"],
+            descriptor["meta"]["simulation"]["name"]
+            .replace("/", "-")
+            .replace("\\", "-"),
         )
         os.makedirs(target_folder, exist_ok=True)
 
         for link in descriptor["links"]:
             if link["rel"] == "lizard-scenario-results":
+                grid = deepcopy(descriptor["meta"]["grid"])
                 results = get_tenant_file_descriptor_view(
                     file["descriptor_id"], "lizard-scenario-results"
                 )
-                result_browser = ResultBrowser(None, results)
+                result_browser = ResultBrowser(None, results, grid["crs"])
                 if result_browser.exec() == QDialog.DialogCode.Accepted:
-                    result_ids = result_browser.get_selected_results_id()
+                    result_ids, nodata, pixelsize, crs = (
+                        result_browser.get_selected_results()
+                    )
                     if len(result_ids) == 0:
                         return
                     filtered_result_ids = []
@@ -416,10 +438,14 @@ class Loader(QObject):
                             filtered_result_ids.append(result_id)
 
                     self.lizard_result_download_worker = LizardResultDownloadWorker(
-                        project,
-                        file,
-                        filtered_result_ids,
-                        target_folder,
+                        project=project,
+                        file=file,
+                        result_ids=filtered_result_ids,
+                        target_folder=target_folder,
+                        grid=grid,
+                        nodata=nodata,
+                        crs=crs,
+                        pixelsize=pixelsize,
                     )
 
                     self.lizard_result_download_worker.finished.connect(
@@ -432,6 +458,8 @@ class Loader(QObject):
                         self.on_file_download_progress
                     )
                     self.lizard_result_download_worker.start()
+                else:
+                    self.download_results_cancelled.emit()
 
     @pyqtSlot(dict, dict)
     def download_file(self, project, file):