+++ rana_qgis_plugin/widgets/result_browser.py
@@ -1,11 +1,16 @@
 from typing import List
 
+from qgis.core import QgsCoordinateReferenceSystem
+from qgis.gui import QgsProjectionSelectionWidget
 from qgis.PyQt.QtCore import Qt
+from qgis.PyQt.QtGui import QDoubleValidator
 from qgis.PyQt.QtWidgets import (
     QDialog,
     QDialogButtonBox,
+    QFormLayout,
     QGridLayout,
     QGroupBox,
+    QLineEdit,
     QTableWidget,
     QTableWidgetItem,
     QVBoxLayout,
@@ -16,7 +21,7 @@ from rana_qgis_plugin.utils import get_filename_from_attachment_url
 
 
 class ResultBrowser(QDialog):
-    def __init__(self, parent, results):
+    def __init__(self, parent, results, scenario_crs):
         super().__init__(parent)
         self.setWindowTitle(PLUGIN_NAME)
         self.setMinimumWidth(400)
@@ -24,19 +29,88 @@ class ResultBrowser(QDialog):
         self.setLayout(layout)
 
         self.selected_results = []
+        self.selected_nodata = None
+        self.selected_pixelsize = None
+        self.selected_crs = None
 
         results_group = QGroupBox("Results", self)
         results_group.setLayout(QGridLayout())
 
-        self.table = QTableWidget(self)
-        results_group.layout().addWidget(self.table)
-        self.table.setColumnCount(2)
-        self.table.verticalHeader().setVisible(False)
-        self.table.horizontalHeader().setStretchLastSection(True)
-        self.table.setHorizontalHeaderLabels(["Type", "File name"])
+        postprocessed_rasters_group = QGroupBox("Generate raster results", self)
+        postprocessed_rasters_group.setLayout(QGridLayout())
+
+        self.results_table = QTableWidget(self)
+        results_group.layout().addWidget(self.results_table)
+        self.results_table.setColumnCount(2)
+        self.results_table.verticalHeader().setVisible(False)
+        self.results_table.horizontalHeader().setStretchLastSection(True)
+        self.results_table.setHorizontalHeaderLabels(["Type", "File name"])
+
+        self.postprocessed_rasters_table = QTableWidget(self)
+        postprocessed_rasters_group.layout().addWidget(self.postprocessed_rasters_table)
+        self.postprocessed_rasters_table.setColumnCount(2)
+        self.postprocessed_rasters_table.verticalHeader().setVisible(False)
+        self.postprocessed_rasters_table.horizontalHeader().setStretchLastSection(True)
+        self.postprocessed_rasters_table.setHorizontalHeaderLabels(
+            ["Type", "File name"]
+        )
+
+        inputs_group = QGroupBox("Generated raster result settings", self)
+        inputs_form = QFormLayout(self)
+        inputs_group.setLayout(inputs_form)
+
+        self.no_data_box = QLineEdit(self)
+        self.pixelsize_box = QLineEdit(self)
+        self.crs_select_box = QgsProjectionSelectionWidget(self)
+
+        self.no_data_box.setValidator(
+            QDoubleValidator(
+                bottom=-1000000.0,
+                top=1000000.0,
+                decimals=1,
+                notation=QDoubleValidator.StandardNotation,
+            )
+        )
+        self.pixelsize_box.setValidator(
+            QDoubleValidator(
+                bottom=0.00001,
+                top=1000000.00000,
+                decimals=5,
+                notation=QDoubleValidator.StandardNotation,
+            )
+        )
+
+        self.no_data_box.setText("-9999.0")
+        self.pixelsize_box.setText("1.00000")
+        self.crs_select_box.setCrs(QgsCoordinateReferenceSystem(scenario_crs))
+
+        inputs_form.addRow("NO DATA value:", self.no_data_box)
+        inputs_form.addRow("Pixel size:", self.pixelsize_box)
+        inputs_form.addRow("CRS:", self.crs_select_box)
+
+        # only show raster download options if raster is selected
+        self.raster_selected = False
+
+        def check_raster_selected():
+            select_states = [
+                self.postprocessed_rasters_table.item(i, 0).checkState()
+                for i in range(self.postprocessed_rasters_table.rowCount())
+            ]
+            if any(state == Qt.CheckState.Checked for state in select_states):
+                self.no_data_box.setEnabled(True)
+                self.pixelsize_box.setEnabled(True)
+                self.crs_select_box.setEnabled(True)
+            else:
+                self.no_data_box.setEnabled(False)
+                self.pixelsize_box.setEnabled(False)
+                self.crs_select_box.setEnabled(False)
+
+        self.postprocessed_rasters_table.cellChanged.connect(check_raster_selected)
+
+        postprocessed_rasters_group.layout().addWidget(inputs_group)
 
         for i, result in enumerate([r for r in results if r["attachment_url"]]):
-            self.table.insertRow(self.table.rowCount())
+            self.results_table.insertRow(self.results_table.rowCount())
             type_item = QTableWidgetItem(result["name"])
             type_item.setFlags(
                 Qt.ItemFlag.ItemIsEnabled | Qt.ItemFlag.ItemIsUserCheckable
@@ -45,13 +119,39 @@ class ResultBrowser(QDialog):
             type_item.setData(Qt.ItemDataRole.UserRole, int(result["id"]))
 
             file_name = get_filename_from_attachment_url(result["attachment_url"])
+
             file_name_item = QTableWidgetItem(file_name)
             file_name_item.setFlags(Qt.ItemFlag.ItemIsEnabled)
-            self.table.setItem(i, 0, type_item)
-            self.table.setItem(i, 1, file_name_item)
+            self.results_table.setItem(i, 0, type_item)
+            self.results_table.setItem(i, 1, file_name_item)
 
-        self.table.resizeColumnsToContents()
+        # timeseries rasters
+        excluded_rasters = ["depth-dtri", "rain-quad", "s1-dtri"]
+
+        for i, result in enumerate(
+            [r for r in results if r["raster_id"] and r["code"] not in excluded_rasters]
+        ):
+            self.postprocessed_rasters_table.insertRow(
+                self.postprocessed_rasters_table.rowCount()
+            )
+            type_item = QTableWidgetItem(result["name"])
+            type_item.setFlags(
+                Qt.ItemFlag.ItemIsEnabled | Qt.ItemFlag.ItemIsUserCheckable
+            )
+            type_item.setCheckState(Qt.CheckState.Unchecked)
+            type_item.setData(Qt.ItemDataRole.UserRole, int(result["id"]))
+
+            file_name = result["code"]
+            file_name_item = QTableWidgetItem(file_name)
+            file_name_item.setFlags(Qt.ItemFlag.ItemIsEnabled)
+            self.postprocessed_rasters_table.setItem(i, 0, type_item)
+            self.postprocessed_rasters_table.setItem(i, 1, file_name_item)
+
+        self.results_table.resizeColumnsToContents()
         layout.addWidget(results_group)
+
+        self.postprocessed_rasters_table.resizeColumnsToContents()
+        layout.addWidget(postprocessed_rasters_group)
 
         buttonBox = QDialogButtonBox(
             QDialogButtonBox.StandardButton.Ok | QDialogButtonBox.StandardButton.Cancel
@@ -60,15 +160,30 @@ class ResultBrowser(QDialog):
         buttonBox.rejected.connect(self.reject)
         layout.addWidget(buttonBox)
 
-    def get_selected_results_id(self) -> List[int]:
-        return self.selected_results
+    def get_selected_results(self) -> List[int]:
+        return (
+            self.selected_results,
+            self.selected_nodata,
+            self.selected_pixelsize,
+            self.selected_crs,
+        )
 
     def accept(self) -> None:
         self.selected_results = []
-        for r in range(self.table.rowCount()):
-            name_item = self.table.item(r, 0)
+        for r in range(self.results_table.rowCount()):
+            name_item = self.results_table.item(r, 0)
             if name_item.checkState() == Qt.CheckState.Checked:
                 id = int(name_item.data(Qt.ItemDataRole.UserRole))
                 self.selected_results.append(id)
+
+        for r in range(self.postprocessed_rasters_table.rowCount()):
+            name_item = self.postprocessed_rasters_table.item(r, 0)
+            if name_item.checkState() == Qt.CheckState.Checked:
+                id = int(name_item.data(Qt.ItemDataRole.UserRole))
+                self.selected_results.append(id)
+
+        self.selected_nodata = float(self.no_data_box.text())
+        self.selected_pixelsize = float(self.pixelsize_box.text())
+        self.selected_crs = self.crs_select_box.crs().authid()
 
         return super().accept()